Add repository for querying the data from our database
The CrudRepository<> implements a lot of crud operations for us. However, those that we need can either be specified like we did with the queries or we can use the built-in queries from CrudRepository within DogService to get all dogs or get a dog by ID, then process the resulting object as well.