Add a GraphQL schema for our Dog model
The schema implements the following query operations:
+ findDogBreeds
+ findDogBreedById
+ findAllDogNames

And the following mutators:
+ deleteDogBreed
+ updateDogName
+ newDog

Note that we did not implement per se all of the queries, but by using GraphQL, the user will be able to specify which fields they want from a query. Thus, by adding queries for find all dogs and find dogs by id, these operations actually exist.

Also, we don't need the service and the controller to implement a GraphQL API. However, if we want to have a REST API available, we can add those files to the application as we are currently doing.