Create Query and Mutation classes based on our GraphQL Schema
To query, we implement a query resolver that will use our DogRepository to query the database. It allows Spring to automatically detect and call the right method in response to one of the GraphQL queries defined inside of the schema. For instance, the findAllDogs query in the schema will be matched to Query.findAllDogs() which retrieves a list of Dogs.

We can also modify the data on the server by using a mutator. Here, it is represented by the Mutation.java class which implements GraphQLMutationResolver which will allow spring to select the right method to call in the repository after the user use a mutation from the schema.