Create custom exceptions for our API and refactor our model and service
The first step to create a custom exception is to extend RuntimeException and then make it specific to GraphQl by implementing GraphQLError. GraphQLError contains a fields called extensions which allows us to pass additional data to the error object sent back to the client. Here, we are passing the id of the invalid Dog. When the message will go back to the client, it will have a generic error message defined in the Mutation class and also the id of the specific location not found.