Design unit test for our controller
@RunWith() at the beginning of the class defines the runner class to be used to run test cases. Here, we use SpringRunner since we use Spring to build the app.
@WebMvcTest() get passed the name of the controller we are trying to test. Here, DogController.
MockMvc offers a powerful way to quickly test MVC controllers without needing to start a full HTTP Server. Note That @WebMvcTest auto configure MockMvc.
We mock the DogService by using @MockBean which creates a mockito mock of the dependencies that are dependencies to the controller.
@Test marks that it is a unit test
mockMvc.perform() simulates an HTTP request to a REST controller. We make a get request to the /dogs/location and /1/breed/.
.andExpect are expectations. They can be set on HTTP responses received from the controller class.
verify() is ued to verify how many times a mock method has been called. Here since we are mocking the dog service, we want it to be called once and the retrieveDogs() and retrieveDogBreedById() to be called.