Add WebTestClient dependency and finish integration testing class for our controller
@SpringBootTest() is mainly used for integration testing and the random port option starts the server with random port to avoid port conflicts in a testing environment.
@AutoConfigureMockMvc() inject spring mock mvc. It simulates calling the code from the client exactly the same way as if we were processing an HTTP request.
Spring Boot automatically provide the TestRestTemplate. All we have to do is to autowire it. It allows us to consume REST APIs in a testing scenario programmatically.

The getForEntity of the rest template is going to make a get request to the specified URL and the response is going to be stored in the response object. We then assert that the response code is 200 (OK)